#52 refactor : 컨트롤 패널 분리
